requirements.txt
fastapi
uvicorn
torch==2.2.0+cpu
torchvision==0.17.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi
pillow
streamlit
requests
python-multipart
scikit-learn
numpy<2